tidy: Fix "multiplication implicit widening" in mythcdrom.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

Fix by forcing the compiler to perform the multiplication as a 64bit
number.

+++ mythtv/libs/libmythbase/mythcdrom.cpp
@@ -178,7 +178,7 @@ static int def_read(udfread_block_input *p_gen, uint32_t lba, void *buf, uint32_
     int result = -1;
     auto *p = (blockInput_t *)p_gen;
 
-    if (p && p->m_file && (p->m_file->Seek(lba * UDF_BLOCK_SIZE, SEEK_SET) != -1))
+    if (p && p->m_file && (p->m_file->Seek(static_cast<long long>(lba) * UDF_BLOCK_SIZE, SEEK_SET) != -1))
         result = p->m_file->Read(buf, nblocks * UDF_BLOCK_SIZE) / UDF_BLOCK_SIZE;
 
     return result;
